requirements.txt
asgiref==3.11.1
Django==6.0.3
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.0
django-filter==25.1
pillow==12.2.0
sqlparse==0.5.5
